+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -1,0 +1,3 @@
+export * from "@/domain/usecases/add-account";
+export * from "@/presentation/protocols";
+export * from "@/presentation/protocols/email-validator";
